allow to drag and drop directly into the chat table view, not only the message input bar
